chore: event-panel-view move tests in tests folder
