+++ src/view/event-panel-view/tests/participants-displayer.test.tsx
@@ -8,13 +8,13 @@ import React from 'react';
 import { combineReducers, configureStore } from '@reduxjs/toolkit';
 import { screen, waitFor } from '@testing-library/react';
 
-import { DisplayedParticipant } from './participants-displayer';
-import * as shell from '../../carbonio-ui-commons/test/mocks/carbonio-shell-ui';
-import defaultSettings from '../../carbonio-ui-commons/test/mocks/settings/default-settings';
-import { setupTest } from '../../carbonio-ui-commons/test/test-setup';
-import { PREFS_DEFAULTS } from '../../constants';
-import * as ParticipantDisplayerAction from '../../store/actions/participant-displayer-actions';
-import { reducers } from '../../store/redux';
+import * as shell from '../../../carbonio-ui-commons/test/mocks/carbonio-shell-ui';
+import defaultSettings from '../../../carbonio-ui-commons/test/mocks/settings/default-settings';
+import { setupTest } from '../../../carbonio-ui-commons/test/test-setup';
+import { PREFS_DEFAULTS } from '../../../constants';
+import * as ParticipantDisplayerAction from '../../../store/actions/participant-displayer-actions';
+import { reducers } from '../../../store/redux';
+import { DisplayedParticipant } from '../participants-displayer';
 
 jest.setTimeout(20000);
